Вынес регистрацию сервисов, middlewares и handlers из Program.cs в отдельные классы
