+++ LoadBalancer/Strategies/Implementations/LeastConnectionsStrategy.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Concurrent;
 using LoadBalancer.Configurations;
+using ILogger = Serilog.ILogger;
 
 namespace LoadBalancer.Strategies.Implementations;
 
@@ -8,10 +9,12 @@ public class LeastConnectionsStrategy :ILoadBalancerStrategy
     private readonly ServerConfig[] _servers;
     private readonly ConcurrentDictionary<string, int> _connectionCounts = new();
     private readonly object _lock = new();
+    private readonly ILogger _logger;
 
-    public LeastConnectionsStrategy(ServerConfig[] servers)
+    public LeastConnectionsStrategy(ServerConfig[] servers, ILogger logger)
     {
         _servers = servers;
+        _logger = logger;
         foreach (var server in servers)
         {
             _connectionCounts[server.Url] = 0;
@@ -25,12 +28,12 @@ public class LeastConnectionsStrategy :ILoadBalancerStrategy
             var targetServer = _servers
                 .OrderBy(s => _connectionCounts[s.Url])
                 .First();
-
+            
             _connectionCounts.AddOrUpdate(
                 targetServer.Url,
                 1,
                 (_, count) => count + 1);
-
+            
             return targetServer;
         }
     }
